LibraryManagementSystem/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-phonenumber-field==8.0.0
phonenumberslite==8.13.51
sqlparse==0.5.2
tzdata==2024.2
